requirements.txt
streamlit
langchain
langchain-groq
langchain-community
tavily-python
python-dotenv
yfinance